+++ windows/runner/main.cpp
@@ -52,6 +52,7 @@ int APIENTRY wWinMain(_In_ HINSTANCE instance, _In_opt_ HINSTANCE prev,
     return EXIT_FAILURE;
   }
   window.SetQuitOnClose(true);
+  window.SetMinimumSize(Win32Window::Size(960, 640));
 
   ::MSG msg;
   while (::GetMessage(&msg, nullptr, 0, 0)) {